Moved the Editor to separate component
